requirements.txt
catkin-pkg==1.0.0
docutils==0.21.2
empy==3.3.4
fastcrc==0.3.2
grpcio==1.74.0
lark==1.2.2
lxml==6.0.0
mavsdk==3.8.0
protobuf==6.31.1
pymavlink==2.4.48
pyparsing==3.2.3
pyserial==3.5
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
shiboken6==6.9.1
six==1.17.0
